+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- MV-275 - Open visio meeting from sheet not in meeting, in a context of multi-sheet user
 
 ## [2.86.2] - 2021-04-02
 ### Fixed